# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "This is an example API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/apprenants/{id}": {
            "get": {
                "tags": [
                    "apprenants"
                ],
                "summary": "Get Apprenant Detail",
                "description": "Get Apprenant Detail",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Apprenant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "apprenants"
                ],
                "summary": "Update apprenant in DB",
                "description": "Update apprenant in DB",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of apprenant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "nom"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "prenom"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "email"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "date_naissance"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "telephone"
                                    },
                                    "cni": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "cni"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "genre"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "apprenants"
                ],
                "summary": "Delete Apprenant",
                "description": "Delete Apprenant",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of Apprenant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/promos": {
            "get": {
                "tags": [
                    "promos"
                ],
                "summary": "Get list of promos",
                "description": "Get list of promos",
                "operationId": " index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order  accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "promos"
                ],
                "summary": "Store promo in DB",
                "description": "Store promo in DB",
                "operationId": " store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "libelle"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "date_debut"
                                    },
                                    "date_fin_prevue": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "date_fin_prevue"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/promos/{id}": {
            "get": {
                "tags": [
                    "promos"
                ],
                "summary": "Get promo Detail",
                "description": "Get promo Detail",
                "operationId": " show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of promo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "promos"
                ],
                "summary": "Update promo in DB",
                "description": "Update promo in DB",
                "operationId": " update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of promo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libelle",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "libelle": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "libelle"
                                    },
                                    "date_debut": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "date_debut"
                                    },
                                    "date_fin_prevue": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "date_fin_prevue"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "promos"
                ],
                "summary": "Delete promo",
                "description": "Delete promo",
                "operationId": " destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of promo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}